+++ test/base/src/jvmTest/kotlin/io/realm/kotlin/test/compiler/list/ListTests.kt
@@ -144,6 +144,7 @@ import io.realm.kotlin.types.RealmInstant
 import io.realm.kotlin.types.ObjectId
 import io.realm.kotlin.types.RealmList
 import io.realm.kotlin.types.RealmObject
+import io.realm.kotlin.types.RealmUUID
 
 import java.lang.Exception
 
@@ -158,6 +159,7 @@ import io.realm.kotlin.types.RealmInstant
 import io.realm.kotlin.types.ObjectId
 import io.realm.kotlin.types.RealmList
 import io.realm.kotlin.types.RealmObject
+import io.realm.kotlin.types.RealmUUID
 
 import java.lang.Exception
 
@@ -172,6 +174,7 @@ import io.realm.kotlin.types.RealmInstant
 import io.realm.kotlin.types.ObjectId
 import io.realm.kotlin.types.RealmList
 import io.realm.kotlin.types.RealmObject
+import io.realm.kotlin.types.RealmUUID
 
 import java.lang.Exception
 
@@ -186,6 +189,7 @@ import io.realm.kotlin.types.RealmInstant
 import io.realm.kotlin.types.ObjectId
 import io.realm.kotlin.types.RealmList
 import io.realm.kotlin.types.RealmObject
+import io.realm.kotlin.types.RealmUUID
 
 import java.lang.Exception
 
@@ -200,6 +204,7 @@ private val UNSUPPORTED_TYPE = """
     import io.realm.kotlin.types.ObjectId
     import io.realm.kotlin.types.RealmList
     import io.realm.kotlin.types.RealmObject
+    import io.realm.kotlin.types.RealmUUID
 
     import java.lang.Exception