+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 * None.
 
 ### Enhancements
+* Added support for `UUID` through a new property type: `RealmUUID`.
 * Support for `Realm.writeCopyTo(configuration)`. 
+* [Sync] Add support for `User.delete()`, making it possible to delete user data on the server side (Issue [#491](https://github.com/realm/realm-kotlin/issues/491)).
 
 ### Fixed
 * `Realm.deleteRealm(config)` would throw an exception if the file didn't exist.
@@ -22,7 +24,7 @@
 * None.
 
 
-## 1.0.2 (YYYY-MM-DD)
+## 1.0.2 (2022-08-05)
 
 ### Breaking Changes
 * None.
@@ -32,7 +34,7 @@
 
 ### Fixed
 * Missing proguard configuration for `CoreErrorUtils`. (Issue [#942](https://github.com/realm/realm-kotlin/issues/942))
-* [Sync] Embedded Objects could not be added to the schema `SyncConfiguration`s. (Issue [#945](https://github.com/realm/realm-kotlin/issues/945)). 
+* [Sync] Embedded Objects could not be added to the schema for `SyncConfiguration`s. (Issue [#945](https://github.com/realm/realm-kotlin/issues/945)). 
 
 ### Compatibility
 * This release is compatible with: